Move organization models to wirecloud.commons
