+++ src/wirecloud/commons/apps.py
@@ -17,4 +17,10 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with Wirecloud.  If not, see <http://www.gnu.org/licenses/>.
 
-default_app_config = 'wirecloud.commons.apps.WirecloudCommonsConfig'
+from django.apps import AppConfig
+
+
+class WirecloudCommonsConfig(AppConfig):
+
+    name = "wirecloud.commons"
+    verbose_name = "WireCloud Core"
