+++ src/wirecloud/platform/fixtures/test_data.json
@@ -87,7 +87,7 @@
     },
     {
         "pk": 1,
-        "model": "platform.organization",
+        "model": "commons.organization",
         "fields": {
             "user": 5,
             "group": 2
@@ -95,7 +95,7 @@
     },
     {
         "pk": 1,
-        "model": "platform.team",
+        "model": "commons.team",
         "fields": {
             "organization": 1,
             "name": "owners",
